增加排产日志及正式排产错误修复调整

+++ APS-Modules/aps-factory-engine/src/main/java/com/zlt/aps/factory/domain/vo/MonthPlanProductionRequirePlanVo.java
@@ -176,8 +176,8 @@ public class MonthPlanProductionRequirePlanVo extends ProductionMonthPlanInit {
      * @param minQty
      * @return
      */
-    public boolean isLess(Long minQty) {
-        if (getHeightProductionQty() > BigDecimal.ZERO.longValue()) {
+    public boolean isLess(Integer minQty) {
+        if (getHeightProductionQty() > BigDecimal.ZERO.intValue()) {
             return getHeightProductionQty() < minQty;
         }
         return getProductionQty() < minQty;
